:recycle: move `replication.Storage` interface to `storage.Streamable` and replace old `Backup` / `Restore` methods with `Streamable` interface
