+++ internal/storage/log/main.go
@@ -6,18 +6,17 @@ import (
 
 	"time"
 
-	"github.com/vladopajic/go-actor/actor"
-
 	"github.com/dgraph-io/badger/v4"
 
 	"link-society.com/flowg/internal/models"
+	"link-society.com/flowg/internal/storage"
+	"link-society.com/flowg/internal/storage/log/transactions"
+	"link-society.com/flowg/internal/utils/kvstore"
 
 	"link-society.com/flowg/internal/utils/ffi/filterdsl"
-	"link-society.com/flowg/internal/utils/kvstore"
-	"link-society.com/flowg/internal/utils/proctree"
-	"link-society.com/flowg/internal/utils/replication"
 
-	"link-society.com/flowg/internal/storage/log/transactions"
+	"github.com/vladopajic/go-actor/actor"
+	"link-society.com/flowg/internal/utils/proctree"
 )
 
 type options struct {
@@ -58,7 +57,7 @@ type Storage struct {
 }
 
 var _ proctree.Process = (*Storage)(nil)
-var _ replication.Storage = (*Storage)(nil)
+var _ storage.Streamable = (*Storage)(nil)
 
 func NewStorage(opts ...func(*options)) *Storage {
 	options := options{
@@ -92,14 +91,6 @@ func NewStorage(opts ...func(*options)) *Storage {
 		Process: process,
 		kvStore: kvStore,
 	}
-}
-
-func (s *Storage) Backup(ctx context.Context, w io.Writer) error {
-	return s.kvStore.Backup(ctx, w, 0)
-}
-
-func (s *Storage) Restore(ctx context.Context, r io.Reader) error {
-	return s.kvStore.Restore(ctx, r)
 }
 
 func (s *Storage) Dump(ctx context.Context, w io.Writer, since uint64) error {
