+++ internal/storage/types.go
@@ -1,11 +1,11 @@
-package replication
+package storage
 
 import (
 	"context"
 	"io"
 )
 
-type Storage interface {
+type Streamable interface {
 	Dump(context.Context, io.Writer, uint64) error
 	Load(context.Context, io.Reader) error
 }
